Made changes based on PR review
